Expose post relationships data in a REST endpoint
